requirements.txt
discord.py
click
asyncpg
